Adding auth details to environment variables and reusing
